+++ src/main/java/pl/sgorski/Tagger/service/AiService.java
@@ -12,6 +12,8 @@ import org.springframework.ai.openai.OpenAiChatOptions;
 import org.springframework.ai.openai.api.ResponseFormat;
 import org.springframework.ai.retry.NonTransientAiException;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.stereotype.Service;
 import pl.sgorski.Tagger.dto.PromptResponse;
 import pl.sgorski.Tagger.exception.AiParsingException;
@@ -25,8 +27,8 @@ public class AiService {
     private String systemContext;
 
     private final OpenAiChatModel chatModel;
-
     private final ObjectMapper objectMapper;
+    private final MessageSource messageSource;
 
     public PromptResponse generateResponse(String userPrompt) {
         var outputConverter = new BeanOutputConverter<>(PromptResponse.class);
@@ -44,16 +46,20 @@ public class AiService {
             try {
                 return objectMapper.readValue(content, PromptResponse.class);
             } catch (Exception ex) {
-                throw new AiParsingException("Failed to parse AI response: " + ex.getMessage());
+                log.error("Failed to parse response: {}", content);
+                String message = messageSource.getMessage("exception.ai.parsing", null, LocaleContextHolder.getLocale());
+                throw new AiParsingException(message);
             }
         } catch (Exception e) {
-            throw new RuntimeException("Failed to execute AI request: " + e.getMessage());
+            String message = messageSource.getMessage("exception.ai.generic", null, LocaleContextHolder.getLocale());
+            log.error("Failed to generate response: {}", e.getMessage());
+            throw new RuntimeException(message);
         }
     }
 
     private Prompt createPrompt(String userPrompt, String jsonSchema) {
         return Prompt.builder()
-                .messages(new SystemMessage(systemContext), new UserMessage(userPrompt))
+                .messages(new SystemMessage(addLanguageHint(systemContext)), new UserMessage(userPrompt))
                 .chatOptions(OpenAiChatOptions.builder()
                         .responseFormat(new ResponseFormat(ResponseFormat.Type.JSON_SCHEMA, jsonSchema))
                         .build()
@@ -67,16 +73,23 @@ public class AiService {
                 
                 Response should be in JSON format with the following format:
                 %s
-                """.formatted(
-                systemContext,
-                jsonSchema
-        );
+                """.formatted(systemContext, jsonSchema);
         return Prompt.builder()
-                .messages(new SystemMessage(schemaHint), new UserMessage(userPrompt))
+                .messages(new SystemMessage(addLanguageHint(schemaHint)), new UserMessage(userPrompt))
                 .build();
     }
 
     private String getAiResponseText(Prompt prompt) {
         return chatModel.call(prompt).getResult().getOutput().getText();
+    }
+
+    private String addLanguageHint(String prompt) {
+        String language = LocaleContextHolder.getLocale().getLanguage();
+        log.info("Adding language hint for language: {}", language);
+        return """
+                %s
+                
+                Respond must be in language: %s
+                """.formatted(prompt, language);
     }
 }